chore(deps): add react-dev-utils resolution (#3105)
