+++ examples/react-native/storybook/Screen/Screen.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { SafeAreaProvider, SafeAreaView } from 'react-native-safe-area-context';
+import { StyleSheet } from 'react-native';
+
+interface ScreenProps {
+  children: React.ReactNode;
+}
+
+export const Screen = ({ children }: ScreenProps) => (
+  <SafeAreaProvider>
+    <SafeAreaView style={styles.container}>{children}</SafeAreaView>
+  </SafeAreaProvider>
+);
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
