+++ examples/react-native/storybook/index.tsx
@@ -7,8 +7,9 @@ import {
 } from '@storybook/react-native';
 import { withKnobs } from '@storybook/addon-knobs';
 import noop from 'lodash/noop';
+
+import { Screen } from './Screen';
 import { loadStories } from './storyLoader';
-import { DefaultContainer } from '@aws-amplify/ui-react-native/src/Authenticator/common';
 
 const STORYBOOK_REQUIRE_CYCLE_PREFIX =
   'Require cycle: node_modules/core-js/internals/microtask.js';
@@ -26,9 +27,9 @@ export function setupStorybook(initStorybook: boolean) {
     // add decorators
     addDecorator(withKnobs);
     addDecorator((Story: any) => (
-      <DefaultContainer>
+      <Screen>
         <Story />
-      </DefaultContainer>
+      </Screen>
     ));
   }
