--- examples/react-native/storybook/ui/Screen.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import { SafeAreaView, StyleSheet } from 'react-native';
-
-interface ScreenProps {
-  children: React.ReactNode;
-}
-
-export const Screen = ({ children }: ScreenProps) => {
-  return <SafeAreaView style={[styles.container]}>{children}</SafeAreaView>;
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
