+++ package.json
@@ -55,6 +55,9 @@
     ]
   },
   "resolutions": {
+    "**/@angular-devkit/build-angular/terser": "5.14.2",
+    "**/@angular-devkit/build-angular/minimatch": "3.0.5",
+    "**/serve/serve-handler/minimatch": "3.0.5",
     "@types/react": "17.0.47",
     "ansi-regex": "5.0.1",
     "ansi-html": "0.0.8",
@@ -72,11 +75,9 @@
     "nth-check": "^2.0.1",
     "prismjs": "^1.25.0",
     "prism-react-renderer": "1.2.1",
+    "react-dev-utils": "^11.0.4",
     "shell-quote": "1.7.3",
     "trim-newlines": "3.0.1",
-    "**/@angular-devkit/build-angular/terser": "5.14.2",
-    "**/@angular-devkit/build-angular/minimatch": "3.0.5",
-    "**/serve/serve-handler/minimatch": "3.0.5",
     "trim": "^0.0.3",
     "vscode-vue-languageservice": "0.27.26",
     "ws": "^7.4.6"
